requirements.txt
dill==0.2.1
Flask==0.10.1
gunicorn==19.9.0
Jinja2==2.7.2
oauth==1.0.1
oauthlib==0.7.2
requests==2.20.0
simplejson==3.8.0
pillow
